+++ shift/admin/home.php
@@ -2,7 +2,6 @@
 	include('../functions.php');
 
 	if (!isAdmin()) {
-		$_SESSION['msg'] = "You must log in first";
 		header('location: ../login.php');
 	}