+++ shift/index.php
@@ -1,9 +1,12 @@
-<?php 
+<?php
 	include('functions.php');
 
 	if (!isLoggedIn()) {
-		$_SESSION['msg'] = "You must log in first";
 		header('location: login.php');
+	}elseif (isMod()) {
+		header('location: moderator/home.php');
+	}elseif (isAdmin()) {
+		header('location: admin/home.php');
 	}
 ?>
 <!DOCTYPE html>
@@ -21,8 +24,8 @@
 		<?php if (isset($_SESSION['success'])) : ?>
 			<div class="error success" >
 				<h3>
-					<?php 
-						echo $_SESSION['success']; 
+					<?php
+						echo $_SESSION['success'];
 						unset($_SESSION['success']);
 					?>
 				</h3>
@@ -37,7 +40,7 @@
 					<strong><?php echo $_SESSION['user']['username']; ?></strong>
 
 					<small>
-						<i  style="color: #888;">(<?php echo ucfirst($_SESSION['user']['user_type']); ?>)</i> 
+						<i  style="color: #888;">(<?php echo ucfirst($_SESSION['user']['user_type']); ?>)</i>
 						<br>
 						<a href="index.php?logout='1'" style="color: red;">logout</a>
 					</small>