+++ shift/login.php
@@ -1,15 +1,17 @@
-<?php include('functions.php') ?>
+<?php
+	include('functions.php')
+?>
 <!DOCTYPE html>
 <html>
 <head>
-	<title>Registration system PHP and MySQL</title>
+	<title>Login</title>
 	<link rel="stylesheet" type="text/css" href="style.css">
 </head>
 <body>
 
 	<div class="header">
 		<h2>NOC JI-NET WORK SCHEDULE</h2>
-		<h2>V0.1.0 <a href="changelog.md" target="_blank">Changelog</a></h2>
+		<h2>V0.1.1 <a href="changelog.md" target="_blank">Changelog</a></h2>
 	</div>
 
 	<form method="post" action="login.php">
@@ -17,7 +19,7 @@
 		<?php echo display_error(); ?>
 
 		<div class="input-group">
-			<label>Username (รหัสพนักงาน)</label>
+			<label>รหัสพนักงาน</label>
 			<input type="text" name="username" >
 		</div>
 		<div class="input-group">
@@ -28,7 +30,7 @@
 			<button type="submit" class="btn" name="login_btn">Login</button>
 		</div>
 		<p>
-			Not yet a member? <a href="register.php">Sign up</a>
+			<!--Not yet a member? <a href="register.php">Sign up</a>-->
 		</p>
 	</form>