+++ shift/moderator/create_user.php
@@ -2,7 +2,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<title>Registration system PHP and MySQL - Create user</title>
+	<title>ADD NEW MEMBER</title>
 	<link rel="stylesheet" type="text/css" href="../style.css">
 	<style>
 		.header {
@@ -23,7 +23,7 @@
 		<?php echo display_error(); ?>
 
 		<div class="input-group">
-			<label>Username</label>
+			<label>รหัสพนักงาน</label>
 			<input type="text" onkeyup="this.value = this.value.toUpperCase();" name="username" value="<?php echo $username; ?>">
 		</div>
 		<div class="input-group">
@@ -33,10 +33,26 @@
 		<div class="input-group">
 			<label>User type</label>
 			<select name="user_type" id="user_type" >
-				<option value=""></option>
+				<option value="user">User</option>
 			<!--	<option value="admin">Admin</option> -->
-			  <option value="user">User</option>
+			  <option value="mod">Moderator</option>
 			</select>
+		</div>
+		<div class="input-group">
+			<label>Shift</label>
+			<select name="shift" id="user_type">
+				<option value="A">A</option>
+			<!--	<option value="admin">Admin</option> -->
+				<option value="B">B</option>
+			</select>
+		</div>
+		<div class="input-group">
+			<label>ชื่อ นามสกุล</label>
+			<input type="text" name="user_name">
+		</div>
+		<div class="input-group">
+			<label>ชื่อเล่น</label>
+			<input type="text" name="user_nickname">
 		</div>
 		<div class="input-group">
 			<label>Password</label>