+++ shift/moderator/home.php
@@ -2,7 +2,6 @@
 	include('../functions.php');
 
 	if (!isMod()) {
-		$_SESSION['msg'] = "You must log in first";
 		header('location: ../login.php');
 	}
 ?>
@@ -48,7 +47,9 @@
 					<small>
 						<i  style="color: #888;">(<?php echo ucfirst($_SESSION['user']['user_type']); ?>)</i>
 						<br>
-						<a href="home.php?logout='1'" style="color: red;">logout</a> &nbsp; <a href="create_user.php"> + add user</a>
+						<a href="home.php?logout='1'" style="color: red;">logout</a>&nbsp;
+						<a href="create_user.php"> + add user</a>&nbsp;
+						<a href="../backend/schedule.php">ตารางงาน</a>
 					</small>
 
 				<?php endif ?>