+++ shift/changelog.md
@@ -1,7 +1,15 @@
+﻿## V0.1.1 (27-01-2020)
+
+### Features
+* **Backend หน้า Update ตารางกะสำหรับ Level Moderator (แต่ยังแก้ไขไม่ได้)
+
+### Bugfix
+* **แก้ไขไม่ให้ Admin/Moderator เมื่อ Login เข้าระบบแล้วสามารถเข้าหน้า Home ของ User ได้โดยเข้าผ่าน Hotlink
+
 ﻿## V0.1.0 (26-01-2020)
 
 ### Features
 * **ระบบ LOGIN/LOGOUT**
 * **ระบบ REGISTER**
-* **สิทธิสมาชิก admin/moderator/user**
+* **สิทธิสมาชิก Admin/Moderator/User**
 * **Frontend/Backend**